requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pillow==11.3.0
python-multipart==0.0.20
tensorflow==2.20.0
numpy==2.3.2